+++ lib/utils.sh
@@ -61,6 +61,7 @@ print_system_table() {
   _row "CPU"     "${RICER_CPU}"
   _row "RAM"     "${RICER_RAM}"
   _row "GPU"     "${RICER_GPU}"
+  _row "Boot"    "${RICER_BOOTLOADER}"
   _row "Arch"    "${RICER_ARCH}"
   echo -e "${CYAN}+${line}+${RESET}"
 }